projects/automatic-content-tagging/streamlit-app/requirements.txt
streamlit
openai
pandas
numpy
python-dotenv